+++ Tests/Common/Integration/IntegrationTestFixtures+UrlLoadBalancing.swift
@@ -18,13 +18,14 @@ extension IntegrationTestFixtures {
         let consensusUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration()
             .validUrlsConsensusUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let fogView = FogViewConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -42,13 +43,14 @@ extension IntegrationTestFixtures {
         let consensusUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration()
             .validUrlsConsensusUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let fogUntrustedTxOut = FogUntrustedTxOutConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -66,13 +68,14 @@ extension IntegrationTestFixtures {
         let consensusUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration()
             .validUrlsConsensusUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let fogMerkleProof = FogMerkleProofConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -90,13 +93,14 @@ extension IntegrationTestFixtures {
         let consensusUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration()
             .validUrlsConsensusUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let fogKeyImage = FogKeyImageConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -114,13 +118,14 @@ extension IntegrationTestFixtures {
         let consensusUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration()
             .validUrlsConsensusUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let fogBlock = FogBlockConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -137,13 +142,14 @@ extension IntegrationTestFixtures {
 
         let fogUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration().validUrlsFogUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let blockchain = BlockchainConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)
@@ -160,13 +166,14 @@ extension IntegrationTestFixtures {
 
         let fogUrlLoadBalancer = try UrlLoadBalancerFixturesIntegration().validUrlsFogUrlBalancer
 
-        let networkConfig = try NetworkConfigFixtures.create(
+        var networkConfig = try NetworkConfigFixtures.create(
             transportProtocol: transportProtocol,
             consensusUrlLoadBalancer: consensusUrlLoadBalancer,
             fogUrlLoadBalancer: fogUrlLoadBalancer)
 
         let consensus = ConsensusConnection(
-            httpFactory: HttpProtocolConnectionFactory(httpRequester: networkConfig.httpRequester),
+            httpFactory: HttpProtocolConnectionFactory(
+                httpRequester: networkConfig.filledHttpRequester()),
             grpcFactory: GrpcProtocolConnectionFactory(),
             config: networkConfig,
             targetQueue: DispatchQueue.main)